hack open the unstructured image in the network
tacking on an command that infinitely tails an unchanging file to keep the main process in the container running. the reason the container will immediately exit on compose is the CMD[bin/bash] at the end of the unstructured dockerfile and subsequently expected -dt tag for an interactive terminal during the docker run command does not exist. No interactive terminal = no main process => process and container exit. A workaround needs to be found (likely an unstructured service that exists outside of the core docker network).